refactored code for IFrame
