+++ src/components/main/stageView/iFrame/hooks/useChangeIframeTheme.ts
@@ -1,0 +1,40 @@
+import {useContext} from 'react';
+
+import { NodeInAppAttribName } from '@_constants/main';
+import { MainContext } from '@_redux/main';
+
+export interface IUseChangeIframThemeProps{
+	contentRef: HTMLIFrameElement | null
+}
+
+export const useChangeIframeTheme = ({contentRef}:IUseChangeIframThemeProps) =>{
+
+	const { validNodeTree, theme } = useContext(MainContext);
+
+	const changeIframeTheme = () => {
+		let uid = "-1";
+		for (let x in validNodeTree) {
+		  if (
+			validNodeTree[x].data.name === "html" &&
+			validNodeTree[x].data.type === "tag"
+		  ) {
+			uid = validNodeTree[x].uid;
+			break;
+		  }
+		}
+		const ele = contentRef?.contentWindow?.document?.querySelector(
+		  `[${NodeInAppAttribName}="${uid}"]`,
+		);
+		if (contentRef) {
+		  if (theme !== "Light") {
+			ele?.setAttribute("data-theme", "dark");
+		  } else {
+			ele?.setAttribute("data-theme", "light");
+		  }
+		}
+	  };
+
+	  return {
+		changeIframeTheme
+	  }
+}
