+++ src/components/main/stageView/iFrame/index.ts
@@ -1,5 +1,3 @@
 import { IFrame } from "./IFrame";
 
 export default IFrame;
-
-export * from "./types";
