--- src/components/main/stageView/iFrame/types.ts
@@ -1,1 +1,0 @@
-export type IFrameProps = {};
