+++ src/components/main/stageView/iFrame/urls/defaultMediaUrls.ts
@@ -1,0 +1,8 @@
+export const defaultImage =
+  "https://user-images.githubusercontent.com/13418616/234660226-dc0cb352-3735-478c-bcc0-d47f73eb3e31.svg";
+
+export const defaultAudio =
+  "https://user-images.githubusercontent.com/13418616/234660225-7195abb2-91e7-402f-aa7d-902bbf7d66f8.svg";
+  
+export const defaultVideo =
+  "https://user-images.githubusercontent.com/13418616/234660227-aeb91595-1ed6-4b46-8197-c6feb7af3718.svg";
